# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Films API",
        "version": "0.2"
    },
    "paths": {
        "/api/signup": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Crée un utilisateur",
                "description": "Crée un nouvel utilisateur avec les données fournies.",
                "operationId": "65b7ef1b1e076c033bf91055f53bd9a1",
                "requestBody": {
                    "description": "Données utilisateur",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "login",
                                    "password",
                                    "email",
                                    "last_name",
                                    "first_name"
                                ],
                                "properties": {
                                    "login": {
                                        "description": "Nom d'utilisateur de l'utilisateur",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Mot de passe de l'utilisateur",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Adresse e-mail de l'utilisateur",
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "last_name": {
                                        "description": "Nom de famille de l'utilisateur",
                                        "type": "string"
                                    },
                                    "first_name": {
                                        "description": "Prénom de l'utilisateur",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Utilisateur créé avec succès"
                    },
                    "422": {
                        "description": "Entité non traitable - Erreur de validation"
                    },
                    "500": {
                        "description": "Erreur de serveur"
                    }
                }
            }
        },
        "/api/signin": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Connexion de l'utilisateur",
                "description": "Authentifie l'utilisateur avec les informations fournies et retourne un jeton d'authentification en cas de succès.",
                "operationId": "b5cee56757c3c79c5b5f115fdc33f460",
                "requestBody": {
                    "description": "Données d'identification de l'utilisateur",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "login",
                                    "password"
                                ],
                                "properties": {
                                    "login": {
                                        "description": "Nom d'utilisateur de l'utilisateur",
                                        "type": "string",
                                        "maxLength": 50
                                    },
                                    "password": {
                                        "description": "Mot de passe de l'utilisateur",
                                        "type": "string",
                                        "maxLength": 255
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Connexion réussie - Un jeton d'authentification est retourné.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "description": "Token d'authentification",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Connexion échouée"
                    },
                    "422": {
                        "description": "Entité non traitable - Erreur de validation"
                    },
                    "500": {
                        "description": "Erreur de serveur"
                    }
                }
            }
        },
        "/api/signout": {
            "get": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Déconnexion de l'utilisateur",
                "description": "Déconnecte l'utilisateur en révoquant tous ses jetons d'authentification.",
                "operationId": "43ce3b3d0b0c7a68db4f95f44ddf0a2c",
                "responses": {
                    "204": {
                        "description": "Déconnexion réussie - Les jetons d'authentification ont été révoqués.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Échec de la déconnexion - Utilisateur non authentifié."
                    },
                    "500": {
                        "description": "Erreur de serveur."
                    }
                },
                "security": [
                    {
                        "Token": []
                    }
                ]
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Récupère un utilisateur par son ID",
                "description": "Récupère les informations d'un utilisateur en fonction de son ID.",
                "operationId": "b3d27f655646f5454083980a35a6b4f5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'utilisateur à récupérer",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Utilisateur récupéré avec succès",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Utilisateur non trouvé - L'ID fourni ne correspond à aucun utilisateur"
                    },
                    "500": {
                        "description": "Erreur de serveur"
                    }
                },
                "security": [
                    {
                        "Token": []
                    }
                ]
            }
        },
        "/api/users/{id}/password": {
            "patch": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Mettre à jour le mot de passe de l'utilisateur",
                "description": "Permet à l'utilisateur de mettre à jour son mot de passe.",
                "operationId": "da4638221a704ca9719a5df92d943473",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Nouveaux mots de passe de l'utilisateur",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "new_password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "new_password": {
                                        "description": "Nouveau mot de passe de l'utilisateur",
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "password_confirmation": {
                                        "description": "Confirmation du nouveau mot de passe",
                                        "type": "string",
                                        "maxLength": 255
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mot de passe mis à jour avec succès",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Données invalides - Le nouveau mot de passe et sa confirmation ne correspondent pas"
                    },
                    "401": {
                        "description": "Non autorisé - Utilisateur non authentifié"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Utilisateur non trouvé - L'ID de l'utilisateur n'existe pas"
                    },
                    "500": {
                        "description": "Erreur de serveur"
                    }
                },
                "security": [
                    {
                        "Token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Token": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}